requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
pytest
httpx
pymongo
python-dotenv